step6

+++ src/ENet.NET/ENetProtocolConst.cs
@@ -9,6 +9,6 @@ public static class ENetProtocolConst
     public const uint ENET_PROTOCOL_MAXIMUM_WINDOW_SIZE = 65536;
     public const uint ENET_PROTOCOL_MINIMUM_CHANNEL_COUNT = 1;
     public const uint ENET_PROTOCOL_MAXIMUM_CHANNEL_COUNT = 255;
-    public const uint ENET_PROTOCOL_MAXIMUM_PEER_ID = 0xFFFF;
+    public const ushort ENET_PROTOCOL_MAXIMUM_PEER_ID = 0xFFFF;
     public const uint ENET_PROTOCOL_MAXIMUM_FRAGMENT_COUNT = 1024 * 1024;
 }
